Wild refactoring.
Remove all graph subpackages and move all the in-memory stuff into the
graph pacakge.

Interfaces are now degined on the main package level
