+++ graph/copy_test.go
@@ -1,4 +1,4 @@
-package memory
+package graph
 
 import (
 	"reflect"
@@ -12,7 +12,7 @@ func TestCopy(t *testing.T) {
 			t.Fatalf("failed to create graph: %v", err)
 		}
 
-		g2 := GraphDeepCopy(g)
+		g2 := DeepCopy(g)
 
 		if !reflect.DeepEqual(g, g2) {
 			t.Fatalf("expected graphs to be equal g: %#v, g2: %#v", g, g2)
@@ -21,7 +21,7 @@ func TestCopy(t *testing.T) {
 
 	t.Run("Non-EmptyGraph", func(t *testing.T) {
 		g := MustGraph(t)
-		g2 := GraphDeepCopy(g)
+		g2 := DeepCopy(g)
 
 		if !reflect.DeepEqual(g, g2) {
 			t.Fatalf("expected graphs to be equal g: %#v, g2: %#v", g, g2)
