+++ graph/marshal/dot/dot.go
@@ -1,9 +1,9 @@
 package dot
 
 import (
-	"gonum.org/v1/gonum/graph/encoding/dot"
+	"github.com/milosgajdos/go-hypher"
 
-	"github.com/milosgajdos/go-hypher/graph"
+	"gonum.org/v1/gonum/graph/encoding/dot"
 )
 
 // Marshaler is used for marshaling graph to DOT format.
@@ -23,6 +23,6 @@ func NewMarshaler(name, prefix, indent string) (*Marshaler, error) {
 }
 
 // Marshal marshal g into DOT and returns it.
-func (m *Marshaler) Marshal(g graph.Graph) ([]byte, error) {
+func (m *Marshaler) Marshal(g hypher.Graph) ([]byte, error) {
 	return dot.Marshal(g, m.name, m.prefix, m.indent)
 }
